Fix "a/**/b" matching "a/bb"
Quoting https://git-scm.com/docs/gitignore#_pattern_format:

  A slash followed by two consecutive asterisks then a slash matches
  zero or more directories. For example, "a/**/b" matches "a/b",
  "a/x/b", "a/x/y/b" and so on.

I read this to mean that the "/**/" pattern should only match at
directory boundaries, i.e. it is NOT meant to be equivalent to "/**".

Fix this by making the last '/' mandatory when the preceding ".*"
regex match is non-empty.

In other words, for "/**/" patterns, change the resulting regex from
  [/].*[/]?
to
  [/](.*[/])?