Fix handling of ***, i.e. to be treated like *
Before this commit, triple-asterisks (or more) are treted as an invalid
pattern, and they cause rule_from_pattern() to return None.

Quoting https://git-scm.com/docs/gitignore#_pattern_format
(after it specifies how double-asterisks are handled):

  Other consecutive asterisks are considered regular asterisks and
  will match according to the previous rules.

I read this to mean that any number of consecutive asterisks (including
double, triple, and more) are to be handled like single asterisks
when not covered by the special cases for double-asterisks.

I don't think the documentation is clear on how to treat triple- (or
more) asterisks in the places where double-asterisks receive special
treatment...