Fix lack of implicit anchoring of patterns to directory separators
A pattern without "*" - like "o.py" - carries an implicit assumption
(not clearly documented in https://git-scm.com/docs/gitignore) that it
should match a complete directory entry, and not freely match within
an entry's name, i.e. it should match "o.py", but not "foo.py" or
"o.pyc".

To accomplish this, ensure that the pattern only matches at the start
of the path, or directly following a directory separator.